backend: Basic SDL3 support

+++ src/worldio.cpp
@@ -336,7 +336,7 @@ void load_world(char* mname) // still supports all map formats that have existed
     settagareas();
     int xs, ys;
     loopi(256) if (texuse) lookuptexture(i, xs, ys);
-    conoutf("read map %s (%d milliseconds)", cgzname, SDL_GetTicks() - lastmillis);
+    conoutf("read map %s (%d milliseconds)", cgzname, (int)(SDL_GetTicks() - lastmillis));
     conoutf("%s", hdr.maptitle);
     startmap(mname);
     loopl(256)
